% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_periscope_continuous.R
\name{scale_pericope_continuous}
\alias{scale_pericope_continuous}
\alias{scale_periscope_continuous}
\title{Periscope Basic Theme for ggplot}
\usage{
scale_periscope_continuous(low_color = "#e2dcea", high_color = "#3d0284")
}
\description{
theme periscope
}
\examples{
ggplot() + scale_periscope_continuous()
}
\author{
Christine Quan
}
