% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_periscope.R
\name{scale_periscope_colors}
\alias{scale_periscope_colors}
\title{Periscope Basic Theme for ggplot}
\usage{
scale_periscope_colors(periscope_colors = rep(c("#6fb8ff", "#ff7555",
  "#59cec5", "#8849a5", "#c1d30c", "#f215a4", "#9480c1", "#e493c7", "#b5ca92"),
  10))
}
\description{
theme periscope
}
\examples{
ggplot() + scale_periscope_colors()
}
\author{
Christine Quan
}
