% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_periscope.R
\name{theme_periscope}
\alias{theme_periscope}
\title{Periscope Basic Theme for ggplot}
\usage{
theme_periscope(base_size = 1, periscope_colors = c("#6fb8ff", "#ff7555",
  "#59cec5", "#8849a5", "#c1d30c", "#f215a4", "#9480c1", "#e493c7", "#b5ca92"),
  palette = RColorBrewer::brewer.pal("Greys", n = 9),
  color.background = palette[2], color.grid.major = palette[3],
  color.axis.text = palette[6], color.axis.title = palette[6],
  color.title = "#2F0035", color.subtitle = "#535368")
}
\description{
theme periscope
}
\examples{
ggplot() + theme_periscope()
}
\author{
Christine Quan
}
